:art: move crm.rs into own file
